requirements.txt
nonebot2>=2.0.0
nonebot-adapter-onebot>=2.2.0
feedparser>=6.0.8
pydantic>=1.10.0
nonebot2[fastapi]>=2.0.0